+++ src/app/create/_components/CreateItem.tsx
@@ -1,6 +1,6 @@
 import { useFormContext } from 'react-hook-form';
 
-import BackIcon from '/public/icons/back.svg';
+import Header from './item/Header';
 import Items from './item/Items';
 import * as styles from './CreateItem.css';
 
@@ -16,19 +16,7 @@ export default function CreateItem({ onBackClick, onSubmitClick }: CreateItemPro
 
   return (
     <div>
-      <div className={styles.header}>
-        <button onClick={onBackClick}>
-          <BackIcon alt="뒤로가기 버튼" />
-        </button>
-        <h1 className={styles.headerTitle}>리스트 생성</h1>
-        <button
-          onClick={onSubmitClick}
-          className={isValid ? styles.headerNextButton : styles.headerNextButtonDisabled}
-          disabled={!isValid ? true : false}
-        >
-          완료
-        </button>
-      </div>
+      <Header isNextActive={isValid} onClickNext={onSubmitClick} />
       <div className={styles.article}>
         <h3 className={styles.label}>
           아이템 추가 <span className={styles.required}>*</span>
